Origin/fix diary

+++ controllers/diary/deleteExercise.js
@@ -1,9 +1,15 @@
 const Exercises = require('../../models/exercisesInDiary');
-const { controllerWrapper } = require('../../utils/index');
+const { controllerWrapper, HttpError } = require('../../utils/index');
 
 // Контролер видалення вправи що містяться в щоденнику користувача за визначену дату
 const deleteExercise = controllerWrapper(async (req, res) => {
   const { exerciseId } = req.params;
+  const { _id } = req.user;
+
+  const exercisestResult = await Exercises.find({ exerciseId, owner: _id });
+  if (exercisestResult.length === 0) {
+    throw new HttpError(404, 'Not found');
+  }
 
   const result = await Exercises.findByIdAndRemove(exerciseId);
   //  Response with the object of exercises
