+++ controllers/diary/deleteProduct.js
@@ -1,12 +1,18 @@
 const Products = require('../../models/productsInDiary');
-const { controllerWrapper } = require('../../utils/index');
+const { controllerWrapper, HttpError } = require('../../utils/index');
 
 // Контролер видалення продукту що містяться в щоденнику користувача за визначену дату
 const deleteProduct = controllerWrapper(async (req, res) => {
   const { productId } = req.params;
+  const { _id } = req.user;
+
+  const productResult = await Products.find({ productId, owner: _id });
+  if (productResult.length === 0) {
+    throw new HttpError(404, 'Not found');
+  }
 
   const result = await Products.findByIdAndRemove(productId);
-  //  Response with the object of exercises
+
   res.status(202).json(result);
 });
